Project files rebuild (module style) - load.js and review.js added
